requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.5
deezer-python==7.0.0